requirements.txt
colorlog==6.6.0
python-dotenv==0.19.2
setuptools==59.1.1
